% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_conflicts.R
\docType{methods}
\name{microbiomeplot_conflicts}
\alias{microbiomeplot_conflicts}
\title{Conflicts between the microbiomeplot and other packages}
\usage{
microbiomeplot_conflicts()
}
\value{
microbiomeplot conflicts
}
\description{
This function lists all the conflicts between packages in the microbiomeplot
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
microbiomeplot_conflicts()
}
