% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_microbiomeplot.R
\docType{methods}
\name{microbiomeplot_logo}
\alias{microbiomeplot_logo}
\title{microbiomeplot_logo}
\usage{
microbiomeplot_logo()
}
\value{
logo
}
\description{
Get the detailed information of microbiomeplot package.
}
\examples{
microbiomeplot_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
