% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_utils.R
\name{microbiomeplot_packages}
\alias{microbiomeplot_packages}
\title{List all packages in the microbiomeplot}
\usage{
microbiomeplot_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include microbiomeplot in the list?}
}
\value{
microbiomeplot packages
}
\description{
List all packages in the microbiomeplot
}
\examples{
microbiomeplot_packages()
}
