% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\alias{plot_boxplot.microbiome_dataset}
\title{Boxplot for OTU abundance}
\usage{
plot_boxplot(
  d,
  x,
  y,
  line = NULL,
  violin = FALSE,
  na.rm = FALSE,
  show.points = TRUE
)

\method{plot_boxplot}{microbiome_dataset}(
  object,
  x,
  y,
  line = NULL,
  violin = FALSE,
  na.rm = FALSE,
  show.points = TRUE
)
}
\arguments{
\item{d}{microbiomedataset object}

\item{x}{the Metadata variable to map to the horizontal axis.}

\item{y}{OTU to map on the vertical axis}

\item{line}{The variable to map on lines}

\item{violin}{Use violin version of the boxplot}

\item{na.rm}{Remove NAs}

\item{show.points}{Include data points in the figure}
}
\value{
a ggplot2 object

a ggplot2 object
}
\description{
boxplots/violn plots for single OTUs against a column in sample_info
}
