% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bubble.R
\name{plot_bubble}
\alias{plot_bubble}
\alias{plot_bubble.microbiome_dataset}
\title{Bubble for Microbiome Composition}
\usage{
plot_bubble(
  object,
  fill = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  theme = "void",
  bubble_scale = 1,
  color_palette = "Set3",
  x = "sample_id",
  extract_intensity_by = "sample_id",
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  title_x = "sample_id",
  title_y = "Relative Abundance",
  ...
)

\method{plot_bubble}{microbiome_dataset}(
  object,
  fill = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  theme = "void",
  bubble_scale = 1,
  color_palette = "Set3",
  x = "sample_id",
  extract_intensity_by = "sample_id",
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  title_x = "sample_id",
  title_y = "Relative Abundance",
  ...
)
}
\arguments{
\item{object}{microbiomedataset object containing taxonomic abundance data.}

\item{fill}{fill for bubble plot}

\item{top_n}{nteger specifying the number of top taxa to display individually.
Taxa beyond this number are grouped as "Other". Default is 10.}

\item{theme}{nteger specifying the number of top taxa to display individually.
Taxa beyond this number are grouped as "Other". Default is 10.}

\item{color_palette}{the color palatte for bubble to be colored by}

\item{x}{sample_id}

\item{extract_intensity_by}{the parameter by which is intensity is extracted by}

\item{what}{what kind of intensity}

\item{na.rm}{remove NAs}

\item{relative}{Relative intensity}

\item{title_x}{title of x-axis}

\item{title_y}{title of y-axis}

\item{...}{}
}
\value{
a ggplot2 object
}
\description{
Bubble plots are another way to visualize compositional microbial community data
(other than a stacked bar chart). In a bubble plot the relative abundance of a species in a
sample is scaled to match the size of the corresponding point.
}
\examples{
plot_bubble(global_patterns)
plot_bubble(global_patterns)
}
