% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hierarchical.R
\name{plot_hierarchical}
\alias{plot_hierarchical}
\title{Hierarchial Clustering}
\usage{
plot_hierarchical(
  object,
  method_dist = c("manhattan", "canberra", "bray", "kulczynski", "gower", "morisita",
    "horn", "mountford", "jaccard", "raup", "binomial", "chao", "altGower", "cao",
    "mahalanobis", "clark", "chisq", "chord", "hellinger", "aitchison",
    "robust.aitchison", "euclidean"),
  binary = FALSE,
  diag = FALSE,
  upper = FALSE,
  na.rm = FALSE,
  theme = c("classic", "bw", "dark", "light", "minimal"),
  method_dend = c("ward.D", "single", "complete", "ward.D2"),
  group = "SampleType",
  type = "rectangle",
  ...
)
}
\arguments{
\item{object}{the microbiomedataset object}

\item{method_dist}{the disimilarity metric method}

\item{binary}{boolean for the calculation of the distnace}

\item{diag}{boolean for the calculation of the distnace}

\item{upper}{boolean for the calculation of the distnace}

\item{na.rm}{boolean for the calculation of the distnace}

\item{theme}{the theme for the plot object}

\item{method_dend}{the method by which the hierarchial clustering to be carried out}

\item{group}{the group parameter for the clusters}

\item{type}{type of plot}

\item{...}{other parameters}
}
\value{
a ggplot2 object
}
\description{
Hierarchial Clustering
}
