% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\alias{plot_PCA.microbiome_dataset}
\title{Plotting a PCA plot for microbiome data}
\usage{
plot_PCA(
  object,
  scale = TRUE,
  grad,
  pc_x = 1,
  pc_y = 2,
  size = 2,
  legend = TRUE,
  title = "PCA Plot for Microbiome Data",
  theme = "minimal",
  legend.position = "bottom",
  scree.plot = FALSE,
  ..
)

\method{plot_PCA}{microbiome_dataset}(
  object,
  scale = TRUE,
  grad,
  pc_x = 1,
  pc_y = 2,
  size = 2,
  legend = TRUE,
  title = "PCA Plot for Microbiome Data",
  theme = "minimal",
  legend.position = "bottom",
  scree.plot = FALSE,
  ..
)
}
\arguments{
\item{object}{microbiome_dataset object}

\item{scale}{scaling of the variables}

\item{grad}{The column in the sample_info that the data_points to be labelled}

\item{pc_x}{Principle Component 1}

\item{pc_y}{Principle Component 2}

\item{size}{size of the labels in the graph}

\item{legend}{legend for the plot}

\item{title}{title of the plot}

\item{theme}{theme of the plot}

\item{legend.position}{position of the legend}

\item{scree.plot}{scree plot of the PCs}

\item{..}{}
}
\description{
Plotting a PCA plot for microbiome data
}
