% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_piechart.R
\name{plot_piechart}
\alias{plot_piechart}
\alias{plot_piechart.microbiome_dataset}
\title{Plotting a Pie chart for the global composition of all samples}
\usage{
plot_piechart(
  object,
  tax_level = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  theme = "minimal",
  ...
)

\method{plot_piechart}{microbiome_dataset}(
  object,
  tax_level = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  top_n = 10,
  theme = "minimal",
  ...
)
}
\arguments{
\item{object}{microbiomedataset object}

\item{tax_level}{the level of taxa to fill}

\item{top_n}{default is 10}

\item{...}{other parameters}
}
\value{
a ggplot2 object

a ggplot2 object
}
\description{
Plotting a Pie chart for the global composition of all samples
}
