% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot.R
\name{plot_barplot}
\alias{plot_barplot}
\alias{plot_barplot.microbiome_dataset}
\title{Barplot based on ggplot2}
\usage{
plot_barplot(
  object,
  fill = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  alluvial = FALSE,
  axis.text.x = ggplot2::element_text(angle = 45, hjust = 1, vjust = 1),
  top_n = 5,
  show.legend,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  re_calculate_relative = FALSE,
  x = "sample_id",
  extract_intensity_by,
  color = "black",
  facet_grid,
  ...
)

\method{plot_barplot}{microbiome_dataset}(
  object,
  fill = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  alluvial = FALSE,
  axis.text.x = ggplot2::element_text(angle = 45, hjust = 1, vjust = 1),
  top_n = 5,
  show.legend,
  what = c("sum_intensity", "mean_intensity", "median_intensity"),
  na.rm = TRUE,
  relative = TRUE,
  re_calculate_relative = FALSE,
  x = "sample_id",
  extract_intensity_by,
  color = "black",
  facet_grid,
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{fill}{fill for barplot}

\item{alluvial}{Add alluvial or not.}

\item{axis.text.x}{axis.text.x from ggplot2}

\item{top_n}{default is 5}

\item{show.legend}{show.legend or not}

\item{what}{which you want to mutate}

\item{na.rm}{na.rm}

\item{relative}{relative intensity or not.}

\item{re_calculate_relative}{re-calculate relative abundance or not.}

\item{x}{x axis}

\item{extract_intensity_by}{extract intensity by what?
must be from sample information.}

\item{color}{color for barplot}

\item{facet_grid}{facet_grid}

\item{...}{other params}
}
\value{
ggplot2 class object
}
\description{
Barplot based on ggplot2
}
\examples{
library(microbiomeplot)
data("global_patterns")
global_patterns \%>\%
  activate_microbiome_dataset(what = "sample_info") \%>\%
  filter(SampleType \%in\% c("Feces", "Mock")) \%>\%
  plot_barplot(fill = "Phylum")

global_patterns \%>\%
  activate_microbiome_dataset(what = "sample_info") \%>\%
  filter(SampleType \%in\% c("Feces", "Mock")) \%>\%
  plot_barplot(
    fill = "Phylum",
    relative = TRUE,
    re_calculate_relative = TRUE,
    show.legend = TRUE
  )

global_patterns \%>\%
  activate_microbiome_dataset(what = "sample_info") \%>\%
  filter(SampleType \%in\% c("Feces", "Mock")) \%>\%
  plot_barplot(fill = "Phylum",
               alluvial = TRUE,
               top_n = 10)

global_patterns \%>\%
  activate_microbiome_dataset(what = "sample_info") \%>\%
  filter(SampleType \%in\% c("Feces", "Mock")) \%>\%
  plot_barplot(
    fill = "Phylum",
    relative = TRUE,
    re_calculate_relative = TRUE,
    show.legend = TRUE,
    alluvial = TRUE,
    top_n = 10
  )

global_patterns \%>\%
  activate_microbiome_dataset(what = "sample_info") \%>\%
  filter(SampleType \%in\% c("Feces", "Mock")) \%>\%
  plot_barplot(
    fill = "Phylum",
    relative = TRUE,
    re_calculate_relative = TRUE,
    show.legend = TRUE,
    alluvial = TRUE,
    top_n = 10,
    facet_grid = "SampleType"
  )
}
