% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\alias{plot_boxplot.microbiome_dataset}
\title{Boxplot for OTU abundance}
\usage{
plot_boxplot(
  object,
  x,
  y,
  line = NULL,
  violin = FALSE,
  na.rm = FALSE,
  show.points = TRUE
)

\method{plot_boxplot}{microbiome_dataset}(
  object,
  x,
  y,
  line = NULL,
  violin = FALSE,
  na.rm = FALSE,
  show.points = TRUE
)
}
\arguments{
\item{object}{A microbiomedataset object.}

\item{x}{The metadata variable to map to the horizontal axis.}

\item{y}{OTU to map on the vertical axis.}

\item{line}{The variable to map on lines.}

\item{violin}{Use violin version of the boxplot (default: FALSE).}

\item{na.rm}{Remove NAs (default: FALSE).}

\item{show.points}{Include data points in the figure (default: TRUE).}
}
\value{
A ggplot2 object.
}
\description{
Boxplots/violin plots for single OTUs against a column in sample_info.
}
\examples{
plot_boxplot(global_patterns, x = "sample_id", y = "165578", violin = TRUE, show.points = FALSE)
# Example for global_patterns data
plot_boxplot(global_patterns, x = "sample_id", y = "165578", violin = TRUE, show.points = FALSE)

}
\references{
Lahti, L., Shetty, S., et al. (2017). microbiome: Tools for microbiome analysis in R.
R package version 1.0.0. URL: https://github.com/microbiome/microbiome

To cite this function in publications, use:
\code{citation("microbiome")}
}
\seealso{
\code{\link{extract_expression_data}}, \code{\link{extract_variable_info}}
}
