% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlepack.R
\name{plot_circlepack}
\alias{plot_circlepack}
\alias{plot_circlepack.microbiome_dataset}
\title{Circlepack plot based on ggplot2}
\usage{
plot_circlepack(
  object,
  label_level = "Class",
  weight = c("intensity", "log_intensity"),
  ...
)

\method{plot_circlepack}{microbiome_dataset}(
  object,
  label_level = "Kingdowm",
  weight = c("log_intensity", "intensity"),
  ...
)
}
\arguments{
\item{object}{microbiome_dataset}

\item{label_level}{label levels}

\item{weight}{intensity or log_intensity}

\item{...}{other params}
}
\value{
ggplot2 class object
}
\description{
Circlepack plot based on ggplot2
}
\examples{
library(microbiomeplot)
data("global_patterns")
global_patterns \%>\%
  activate_microbiome_dataset(what = "variable_info") \%>\%
  filter(Kingdom == "Archaea") \%>\%
  plot_circlepack(label_level = "Class")
}
