% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\alias{plot_PCA.microbiome_dataset}
\title{Plotting a PCA plot for the microbiome data}
\usage{
plot_PCA(
  object,
  scale = TRUE,
  colour,
  pc_x = 1,
  pc_y = 2,
  size,
  legend = TRUE,
  title,
  theme,
  legend.position = "bottom",
  scree.plot = TRUE,
  ..
)

\method{plot_PCA}{microbiome_dataset}(
  object,
  scale = TRUE,
  colour,
  pc_x = 1,
  pc_y = 2,
  size,
  legend = TRUE,
  title,
  theme,
  legend.position = "bottom",
  scree.plot = TRUE,
  ..
)
}
\arguments{
\item{object}{microbiome_dataset object}

\item{scale}{scaling parameter for prcomp object}

\item{colour}{color the data points by a column on the sample_info}

\item{pc_x}{Specify the x axis principal component}

\item{pc_y}{Specify the y axis principal component}

\item{..}{other parameters}
}
\value{
a ggplot2 object
}
\description{
Plotting a PCA plot for the microbiome data
}
