% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tSNE.R
\name{plot_tSNE}
\alias{plot_tSNE}
\alias{plot_tSNE.microbiome_dataset}
\title{Plotting a t-SNE plot for visualizing high-dimensional microbiome data}
\usage{
plot_tSNE(
  object,
  fill,
  dims = 2,
  perplexity = 30,
  number_of_iteration = 1000,
  initial_dims = 50,
  pca = TRUE,
  legend = TRUE,
  title = "t-SNE plot for Microbiome Data",
  x_title = "axis 1",
  y_title = "axis 2",
  theme = "minimal",
  legend.position = "bottom",
  size = 3,
  title_size = 10,
  ..
)

\method{plot_tSNE}{microbiome_dataset}(
  object,
  fill,
  dims = 2,
  perplexity = 30,
  number_of_iters = 1000,
  initial_dims = 50,
  pca = TRUE,
  legend = TRUE,
  title = "t-SNE plot for Microbiome Data",
  x_title,
  y_title,
  theme = "minimal",
  legend.position = "bottom",
  size = 3,
  title_size = 10,
  ..
)
}
\arguments{
\item{object}{the microbiomedataset object}

\item{legend}{legend for the plot}

\item{title}{Title of the plot}

\item{theme}{theme for the ggplot}

\item{legend.position}{position of the legend}

\item{size}{size of the labels}

\item{..}{}
}
\value{
A ggplo2 class object
}
\description{
Plotting a t-SNE plot for visualizing high-dimensional microbiome data
}
