% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tSNE.R
\name{plot_tSNE}
\alias{plot_tSNE}
\alias{plot_tSNE.microbiome_dataset}
\title{Plotting a t-SNE plot for visualizing high-dimensional microbiome data}
\usage{
plot_tSNE(
  object,
  grad,
  dims = 2,
  perplexity = 30,
  number_of_iteration = 1000,
  initial_dims = 50,
  pca = TRUE,
  legend = TRUE,
  title = "t-SNE plot for Microbiome Data",
  x_title = "dimension 1",
  y_title = "dimension 2",
  z_title = "dimension 3",
  theme = "minimal",
  legend.position = "bottom",
  size = 3,
  title_size = 10,
  ..
)

\method{plot_tSNE}{microbiome_dataset}(
  object,
  grad,
  dims = 2,
  perplexity = 30,
  number_of_iters = 1000,
  initial_dims = 50,
  pca = TRUE,
  legend = TRUE,
  title = "t-SNE plot for Microbiome Data",
  x_title = "dimension 1",
  y_title = "dimension 2",
  z_title = "dimension 3",
  theme = "minimal",
  legend.position = "bottom",
  size = 3,
  title_size = 10,
  ..
)
}
\arguments{
\item{object}{the microbiomedataset object}

\item{grad}{the column of the sample_info to be colored by}

\item{dims}{the output dimensionality of the t-sne object}

\item{perplexity}{A parameter that controls the balance between local and global aspects of the data when t-SNE is creating the low-dimensional map.}

\item{number_of_iteration}{Number of iterations (default: 1000)}

\item{initial_dims}{e number of dimensions that should be retained in the initial PCA step}

\item{pca}{ogical; Whether an initial PCA step should be performed}

\item{legend}{legend for the plot}

\item{title}{Title of the plot}

\item{x_title}{The x-axis title of the plot}

\item{y_title}{the y-axis title of the plot}

\item{z_title}{the z-axis title of the plot}

\item{theme}{theme for the ggplot}

\item{legend.position}{position of the legend}

\item{size}{size of the labels}

\item{title_size}{the size of the text for the title}

\item{..}{other parameters}
}
\value{
A ggplo2 class object
}
\description{
Plotting a t-SNE plot for visualizing high-dimensional microbiome data
}
